chore(suite-native): Allow to pass Preloaded state to E2E
